+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Show definition in list view with configurable truncation.
   [csenger]
 
+- Rename 'Term' to 'GlossaryTerm'.
+  [timo]
+
 
 1.0.0a3 (2019-05-28)
 --------------------